Moved gpu octree tests to test/gpu.
Added to tests.
Make all test compile.
