+++ test/CMakeLists.txt
@@ -34,6 +34,9 @@ add_subdirectory(common)
 add_subdirectory(features)
 add_subdirectory(filters)
 add_subdirectory(geometry)
+if(HAVE_CUDA)
+  add_subdirectory(gpu)
+endif()
 add_subdirectory(io)
 add_subdirectory(kdtree)
 add_subdirectory(keypoints)
