+++ test/gpu/octree/test_bfrs_gpu.cpp
@@ -98,3 +98,13 @@ TEST (PCL_GPU, bruteForceRadiusSeachGPU)
     std::cout << "avg_result_size = " << avg_size << std::endl;
     ASSERT_GT(avg_size, 5);    
 }
+
+/* ---[ */
+int
+main (int argc, char** argv)
+{
+  testing::InitGoogleTest (&argc, argv);
+  return (RUN_ALL_TESTS ());
+}
+/* ]--- */
+
