+++ test/gpu/octree/test_host_radius_search.cpp
@@ -141,3 +141,4 @@ int main (int argc, char** argv)
     testing::InitGoogleTest (&argc, argv);
     return (RUN_ALL_TESTS ());
 }
+
